feat: Add scripts for collecting and visualizing ablation and benchmarking results
- Implemented `collect_ablation_results.py` to parse and aggregate results from ablation experiments, outputting per-experiment CSVs and a summary.
- Created `collect_benchmarking_results.py` to scan benchmarking results, generating per-layer CSVs and summary statistics.
- Developed `draw_benchmarking.py` for visualizing benchmarking metrics with adaptive y-axis scaling and custom styles.
- Enhanced `train_sae_online.py` with CUDA library preloading, early stopping functionality, and improved logging.
- Updated logging utility to accept optional name parameter.
- Refined intervention functions in `intervention.py` for better compatibility with NNsight.
- Improved trainer metrics handling in `trainer.py`, including padding filtering for activations.
- Adjusted metrics calculations in `ce_loss.py`, `logitlens.py`, and `variance_explained.py` for better tensor handling.
- Enhanced SAE model configuration and training logic in `sae.py`, including anchor loss computation frequency.

+++ src/vasae/metrics/ce_loss.py
@@ -60,7 +60,7 @@ class CELossRecovered(IMetric):
         # CE(sae): SAE reconstruction
         logits_sae = patch_and_forward(
             self.model, input_ids, attention_mask,
-            self.layer_idx, lambda h: sae_model(h).hidden_states_recon,
+            self.layer_idx, lambda h: sae_model(h.float()).hidden_states_recon.to(h.dtype),
         )
         ce_sae = cross_entropy(logits_sae, input_ids, attention_mask)
